+++ packages/web-components/fast-element/src/binding/signal.ts
@@ -3,18 +3,27 @@ import type {
     ExpressionController,
     ExpressionObserver,
 } from "../observation/observable.js";
-import { isString, noop } from "../interfaces.js";
+import { isString } from "../interfaces.js";
 import type { Subscriber } from "../observation/notifier.js";
 import type { DOMPolicy } from "../dom.js";
-import type { HTMLBindingDirective } from "./binding.js";
-import { Binding } from "./html-directive.js";
+import { makeSerializationNoop } from "../platform.js";
+import { Binding } from "./binding.js";
 
 const subscribers: Record<
     string,
     undefined | Subscriber | Set<Subscriber>
 > = Object.create(null);
 
+/**
+ * The gateway to signal APIs.
+ * @public
+ */
 export const Signal = Object.freeze({
+    /**
+     * Subscribes to a signal.
+     * @param signal The signal to subscribe to.
+     * @param subscriber The subscriber.
+     */
     subscribe(signal: string, subscriber: Subscriber) {
         const found = subscribers[signal];
 
@@ -27,6 +36,11 @@ export const Signal = Object.freeze({
         }
     },
 
+    /**
+     * Unsubscribes from the signal.
+     * @param signal The signal to unsubscribe from.
+     * @param subscriber The subscriber.
+     */
     unsubscribe(signal: string, subscriber: Subscriber) {
         const found = subscribers[signal];
 
@@ -38,9 +52,8 @@ export const Signal = Object.freeze({
     },
 
     /**
-     * Sends the specified signal to signaled bindings.
+     * Sends the specified signal to subscribers.
      * @param signal - The signal to send.
-     * @public
      */
     send(signal: string): void {
         const found = subscribers[signal];
@@ -79,12 +92,6 @@ class SignalObserver<TSource = any, TReturn = any, TParent = any> implements Sub
         this.subscriber.handleChange(this.dataBinding.evaluate, this);
     }
 
-    /**
-     * Opts out of JSON stringification.
-     * @internal
-     */
-    toJSON = noop;
-
     private getSignal(controller: ExpressionController<TSource, TParent>): string {
         const options = this.dataBinding.options;
         return isString(options)
@@ -93,13 +100,14 @@ class SignalObserver<TSource = any, TReturn = any, TParent = any> implements Sub
     }
 }
 
+makeSerializationNoop(SignalObserver);
+
 class SignalBinding<TSource = any, TReturn = any, TParent = any> extends Binding<
     TSource,
     TReturn,
     TParent
 > {
     createObserver(
-        directive: HTMLBindingDirective,
         subscriber: Subscriber
     ): ExpressionObserver<TSource, TReturn, TParent> {
         return new SignalObserver(this, subscriber);