+++ packages/web-components/fast-element/src/interfaces.ts
@@ -50,15 +50,25 @@ export type TrustedTypesPolicy = {
 
 /**
  * Reverses all readonly members, making them mutable.
- * @internal
+ * @public
  */
 export type Mutable<T> = {
     -readonly [P in keyof T]: T[P];
 };
 
 /**
+ * A temporary type as a workaround for the TS compiler's erroneous built-in ParameterDecorator type.
+ * @public
+ */
+export type ParameterDecorator = (
+    target: Object,
+    propertyKey: string | undefined,
+    parameterIndex: number
+) => void;
+
+/**
  * The FAST global.
- * @internal
+ * @public
  */
 export interface FASTGlobal {
     /**
@@ -174,6 +184,7 @@ export const enum Message {
     directCallToHTMLTagNotAllowed = 1206,
     onlySetTemplatePolicyOnce = 1207,
     cannotSetTemplatePolicyAfterCompilation = 1208,
+    blockedByDOMPolicy = 1209,
     // 1301 - 1400 Styles
     // 1401 - 1500 Components
     missingElementDefinition = 1401,
@@ -196,19 +207,19 @@ export const enum Message {
 
 /**
  * Determines whether or not an object is a function.
- * @internal
+ * @public
  */
 export const isFunction = (object: any): object is Function =>
     typeof object === "function";
 
 /**
  * Determines whether or not an object is a string.
- * @internal
+ * @public
  */
 export const isString = (object: any): object is string => typeof object === "string";
 
 /**
  * A function which does nothing.
- * @internal
+ * @public
  */
 export const noop = () => void 0;