+++ packages/web-components/fast-foundation/src/divider/divider.options.ts
@@ -1,3 +1,6 @@
+import { Orientation } from "@microsoft/fast-web-utilities";
+import type { ValuesOf } from "../utilities/index.js";
+
 /**
  * Divider roles
  * @public
@@ -18,4 +21,16 @@ export const DividerRole = {
  * The types for Divider roles
  * @public
  */
-export type DividerRole = typeof DividerRole[keyof typeof DividerRole];
+export type DividerRole = ValuesOf<typeof DividerRole>;
+
+/**
+ * Divider orientation
+ * @public
+ */
+export const DividerOrientation = Orientation;
+
+/**
+ * The types for Divider orientation
+ * @public
+ */
+export type DividerOrientation = ValuesOf<typeof DividerOrientation>;