+++ packages/web-components/fast-foundation/src/divider/divider.ts
@@ -1,8 +1,5 @@
 import { attr, FASTElement } from "@microsoft/fast-element";
-import { Orientation } from "@microsoft/fast-web-utilities";
-import { DividerRole } from "./divider.options.js";
-
-export { DividerRole };
+import { DividerOrientation, DividerRole } from "./divider.options.js";
 
 /**
  * A Divider Custom HTML Element.
@@ -29,5 +26,5 @@ export class FASTDivider extends FASTElement {
      * HTML Attribute: orientation
      */
     @attr
-    public orientation: Orientation = Orientation.horizontal;
+    public orientation: DividerOrientation = DividerOrientation.horizontal;
 }