+++ packages/web-components/fast-foundation/src/menu-item/menu-item.ts
@@ -1,13 +1,6 @@
 import type { Placement } from "@floating-ui/dom";
 import { autoUpdate, computePosition, flip, shift, size } from "@floating-ui/dom";
-import {
-    attr,
-    DangerousHTMLDirective,
-    FASTElement,
-    observable,
-    SyntheticViewTemplate,
-    Updates,
-} from "@microsoft/fast-element";
+import { attr, FASTElement, observable, Updates } from "@microsoft/fast-element";
 import {
     keyArrowLeft,
     keyArrowRight,
@@ -15,6 +8,7 @@ import {
     keyEscape,
     keySpace,
 } from "@microsoft/fast-web-utilities";
+import type { StaticallyComposableHTML } from "../utilities/template-helpers.js";
 import type { StartEndOptions } from "../patterns/start-end.js";
 import { StartEnd } from "../patterns/start-end.js";
 import { applyMixins } from "../utilities/apply-mixins.js";
@@ -26,10 +20,10 @@ export { MenuItemRole, roleForMenuItem };
  * Menu Item configuration options
  * @public
  */
-export type MenuItemOptions = StartEndOptions & {
-    checkboxIndicator?: DangerousHTMLDirective | SyntheticViewTemplate;
-    expandCollapseGlyph?: DangerousHTMLDirective | SyntheticViewTemplate;
-    radioIndicator?: DangerousHTMLDirective | SyntheticViewTemplate;
+export type MenuItemOptions = StartEndOptions<FASTMenuItem> & {
+    checkboxIndicator?: StaticallyComposableHTML<FASTMenuItem>;
+    expandCollapseGlyph?: StaticallyComposableHTML<FASTMenuItem>;
+    radioIndicator?: StaticallyComposableHTML<FASTMenuItem>;
 };
 
 /**
@@ -196,7 +190,9 @@ export class FASTMenuItem extends FASTElement {
 
             case keyArrowRight:
                 //open/focus on submenu
-                this.expandAndFocus();
+                this.expanded && this.submenu
+                    ? this.submenu.focus()
+                    : this.expandAndFocus();
                 return false;
 
             case keyEscape:
@@ -237,6 +233,7 @@ export class FASTMenuItem extends FASTElement {
         if (!this.focusSubmenuOnLoad) {
             return;
         }
+
         this.focusSubmenuOnLoad = false;
         if (this.submenu) {
             this.submenu.focus();
@@ -286,6 +283,7 @@ export class FASTMenuItem extends FASTElement {
         if (!this.hasSubmenu) {
             return;
         }
+
         this.focusSubmenuOnLoad = true;
         this.expanded = true;
     };