+++ packages/web-components/fast-foundation/src/slider/slider.ts
@@ -166,7 +166,7 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
      */
     @attr({ converter: nullableNumberConverter })
     public min: number = 0; // Map to proxy element.
-    private minChanged(): void {
+    protected minChanged(): void {
         if (this.proxy instanceof HTMLInputElement) {
             this.proxy.min = `${this.min}`;
         }
@@ -184,7 +184,7 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
      */
     @attr({ converter: nullableNumberConverter })
     public max: number = 10; // Map to proxy element.
-    private maxChanged(): void {
+    protected maxChanged(): void {
         if (this.proxy instanceof HTMLInputElement) {
             this.proxy.max = `${this.max}`;
         }
@@ -199,8 +199,8 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
      * HTML Attribute: step
      */
     @attr({ converter: nullableNumberConverter })
-    public step: number = 1; // Map to proxy element.
-    private stepChanged(): void {
+    public step: number | undefined;
+    protected stepChanged(): void {
         if (this.proxy instanceof HTMLInputElement) {
             this.proxy.step = `${this.step}`;
         }
@@ -218,7 +218,7 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
      */
     @attr
     public orientation: Orientation = Orientation.horizontal;
-    private orientationChanged(): void {
+    protected orientationChanged(): void {
         if (this.$fastController.isConnected) {
             this.setThumbPositionForOrientation(this.direction);
         }
@@ -265,8 +265,8 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
     public increment(): void {
         const newVal: number =
             this.direction !== Direction.rtl && this.orientation !== Orientation.vertical
-                ? Number(this.value) + Number(this.step)
-                : Number(this.value) - Number(this.step);
+                ? Number(this.value) + Number(this.stepValue)
+                : Number(this.value) + Number(this.stepValue);
         const incrementedVal: number = this.convertToConstrainedValue(newVal);
         const incrementedValString: string =
             incrementedVal < Number(this.max) ? `${incrementedVal}` : `${this.max}`;
@@ -281,8 +281,8 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
     public decrement(): void {
         const newVal =
             this.direction !== Direction.rtl && this.orientation !== Orientation.vertical
-                ? Number(this.value) - Number(this.step)
-                : Number(this.value) + Number(this.step);
+                ? Number(this.value) - Number(this.stepValue)
+                : Number(this.value) - Number(this.stepValue);
         const decrementedVal: number = this.convertToConstrainedValue(newVal);
         const decrementedValString: string =
             decrementedVal > Number(this.min) ? `${decrementedVal}` : `${this.min}`;
@@ -290,16 +290,20 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
     }
 
     protected keypressHandler = (e: KeyboardEvent) => {
-        if (this.readOnly) {
+        if (this.readOnly || this.disabled) {
             return;
         }
 
         if (e.key === keyHome) {
             e.preventDefault();
-            this.value = `${this.min}`;
+            this.direction !== Direction.rtl && this.orientation !== Orientation.vertical
+                ? (this.value = `${this.min}`)
+                : (this.value = `${this.max}`);
         } else if (e.key === keyEnd) {
             e.preventDefault();
-            this.value = `${this.max}`;
+            this.direction !== Direction.rtl && this.orientation !== Orientation.vertical
+                ? (this.value = `${this.max}`)
+                : (this.value = `${this.min}`);
         } else if (!e.shiftKey) {
             switch (e.key) {
                 case keyArrowRight:
@@ -315,6 +319,14 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
             }
         }
     };
+
+    /**
+     * Gets the actual step value for the slider
+     *
+     */
+    private get stepValue(): number {
+        return this.step === undefined ? 1 : this.step;
+    }
 
     /**
      * Places the thumb based on the current value
@@ -336,8 +348,8 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
                 : `right: ${percentage}%; transition: all 0.2s ease;`;
         } else {
             this.position = this.isDragging
-                ? `bottom: ${percentage}%; transition: none;`
-                : `bottom: ${percentage}%; transition: all 0.2s ease;`;
+                ? `top: ${percentage}%; transition: none;`
+                : `top: ${percentage}%; transition: all 0.2s ease;`;
         }
     }
 
@@ -346,8 +358,8 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
      * are not whole numbers
      */
     private updateStepMultiplier(): void {
-        const stepString: string = this.step + "";
-        const decimalPlacesOfStep: number = !!(this.step % 1)
+        const stepString: string = this.stepValue + "";
+        const decimalPlacesOfStep: number = !!(this.stepValue % 1)
             ? stepString.length - stepString.indexOf(".") - 1
             : 0;
         this.stepMultiplier = Math.pow(10, decimalPlacesOfStep);
@@ -357,8 +369,8 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
         const clientRect: DOMRect = this.track.getBoundingClientRect();
         this.trackWidth = this.track.clientWidth;
         this.trackMinWidth = this.track.clientLeft;
-        this.trackHeight = clientRect.bottom;
-        this.trackMinHeight = clientRect.top;
+        this.trackHeight = clientRect.top;
+        this.trackMinHeight = clientRect.bottom;
         this.trackLeft = this.getBoundingClientRect().left;
         if (this.trackWidth === 0) {
             this.trackWidth = 1;
@@ -410,12 +422,6 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
      *  If the event handler is null it removes the events
      */
     private handleThumbMouseDown = (event: MouseEvent | null): void => {
-        if (event) {
-            if (this.readOnly || this.disabled || event.defaultPrevented) {
-                return;
-            }
-            (event.target as HTMLElement).focus();
-        }
         const eventAction = `${event !== null ? "add" : "remove"}EventListener`;
         window[eventAction]("mouseup", this.handleWindowMouseUp);
         window[eventAction]("mousemove", this.handleMouseMove, { passive: true });
@@ -431,7 +437,6 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
         if (this.readOnly || this.disabled || e.defaultPrevented) {
             return;
         }
-
         // update the value based on current position
         const sourceEvent =
             window.TouchEvent && e instanceof TouchEvent
@@ -496,9 +501,7 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
             window[eventAction]("mousemove", this.handleMouseMove);
 
             if (e) {
-                e.preventDefault();
                 this.setupTrackConstraints();
-                (e.target as HTMLElement).focus();
                 const controlValue: number =
                     this.orientation === Orientation.horizontal
                         ? e.pageX - document.documentElement.scrollLeft - this.trackLeft
@@ -521,15 +524,17 @@ export class FASTSlider extends FormAssociatedSlider implements SliderConfigurat
          * integer and then dividing it to get back to the correct number.
          */
         let constrainedValue: number = value - this.min;
-        const roundedConstrainedValue: number = Math.round(constrainedValue / this.step);
+        const roundedConstrainedValue: number = Math.round(
+            constrainedValue / this.stepValue
+        );
         const remainderValue: number =
             constrainedValue -
-            (roundedConstrainedValue * (this.stepMultiplier * this.step)) /
+            (roundedConstrainedValue * (this.stepMultiplier * this.stepValue)) /
                 this.stepMultiplier;
 
         constrainedValue =
-            remainderValue >= Number(this.step) / 2
-                ? constrainedValue - remainderValue + Number(this.step)
+            remainderValue >= Number(this.stepValue) / 2
+                ? constrainedValue - remainderValue + Number(this.stepValue)
                 : constrainedValue - remainderValue;
         return constrainedValue + this.min;
     }