+++ packages/web-components/fast-foundation/src/tabs/tabs.ts
@@ -9,29 +9,9 @@ import {
     limit,
     uniqueId,
 } from "@microsoft/fast-web-utilities";
-import { StartEnd, StartEndOptions } from "../patterns/index.js";
+import { StartEnd } from "../patterns/start-end.js";
 import { applyMixins } from "../utilities/apply-mixins.js";
-
-/**
- * Tabs option configuration options
- * @public
- */
-export type TabsOptions = StartEndOptions;
-
-/**
- * The orientation of the {@link @microsoft/fast-foundation#(FASTTabs:class)} component
- * @public
- */
-export const TabsOrientation = {
-    vertical: "vertical",
-    horizontal: "horizontal",
-} as const;
-
-/**
- * The types for the Tabs component
- * @public
- */
-export type TabsOrientation = typeof TabsOrientation[keyof typeof TabsOrientation];
+import { TabsOrientation } from "./tabs.options.js";
 
 /**
  * A Tabs Custom HTML Element.
@@ -42,7 +22,6 @@ export type TabsOrientation = typeof TabsOrientation[keyof typeof TabsOrientatio
  * @slot tab - The slot for tabs
  * @slot tabpanel - The slot for tabpanels
  * @csspart tablist - The element wrapping for the tabs
- * @csspart activeIndicator - The visual indicator
  * @fires change - Fires a custom 'change' event when a tab is clicked or during keyboard navigation
  *
  * @public
@@ -63,7 +42,6 @@ export class FASTTabs extends FASTElement {
         if (this.$fastController.isConnected) {
             this.setTabs();
             this.setTabPanels();
-            this.handleActiveIndicatorPosition();
         }
     }
     /**
@@ -88,7 +66,6 @@ export class FASTTabs extends FASTElement {
             );
             this.setTabs();
             this.setTabPanels();
-            this.handleActiveIndicatorPosition();
         }
     }
 
@@ -110,7 +87,6 @@ export class FASTTabs extends FASTElement {
 
             this.setTabs();
             this.setTabPanels();
-            this.handleActiveIndicatorPosition();
         }
     }
 
@@ -132,30 +108,8 @@ export class FASTTabs extends FASTElement {
 
             this.setTabs();
             this.setTabPanels();
-            this.handleActiveIndicatorPosition();
         }
     }
-
-    /**
-     * Whether or not to show the active indicator
-     * @public
-     * @remarks
-     * HTML Attribute: activeindicator
-     */
-    @attr({ attribute: "hide-active-indicator", mode: "boolean" })
-    public hideActiveIndicator = false;
-
-    /**
-     * @internal
-     */
-    @observable
-    public activeIndicatorRef: HTMLElement;
-
-    /**
-     * @internal
-     */
-    @observable
-    public showActiveIndicator: boolean = true;
 
     /**
      * A reference to the active tab
@@ -165,7 +119,6 @@ export class FASTTabs extends FASTElement {
 
     private prevActiveTabIndex: number = 0;
     private activeTabIndex: number = 0;
-    private ticking: boolean = false;
     private tabIds: Array<string>;
     private tabpanelIds: Array<string>;
 
@@ -192,7 +145,7 @@ export class FASTTabs extends FASTElement {
         }
     }
 
-    private setTabs = (): void => {
+    protected setTabs = (): void => {
         const gridHorizontalProperty: string = "gridColumn";
         const gridVerticalProperty: string = "gridRow";
         const gridProperty: string = this.isHorizontal()
@@ -200,14 +153,12 @@ export class FASTTabs extends FASTElement {
             : gridVerticalProperty;
 
         this.activeTabIndex = this.getActiveIndex();
-        this.showActiveIndicator = false;
+
         this.tabs.forEach((tab: HTMLElement, index: number) => {
             if (tab.slot === "tab") {
                 const isActiveTab =
                     this.activeTabIndex === index && this.isFocusableElement(tab);
-                if (!this.hideActiveIndicator && this.isFocusableElement(tab)) {
-                    this.showActiveIndicator = true;
-                }
+
                 const tabId: string = this.tabIds[index];
                 const tabpanelId: string = this.tabpanelIds[index];
                 tab.setAttribute("id", tabId);
@@ -312,44 +263,6 @@ export class FASTTabs extends FASTElement {
                 break;
         }
     };
-
-    private handleActiveIndicatorPosition() {
-        // Ignore if we click twice on the same tab
-        if (
-            this.showActiveIndicator &&
-            !this.hideActiveIndicator &&
-            this.activeTabIndex !== this.prevActiveTabIndex
-        ) {
-            if (this.ticking) {
-                this.ticking = false;
-            } else {
-                this.ticking = true;
-                this.animateActiveIndicator();
-            }
-        }
-    }
-
-    private animateActiveIndicator(): void {
-        this.ticking = true;
-        const gridProperty: string = this.isHorizontal() ? "gridColumn" : "gridRow";
-        const translateProperty: string = this.isHorizontal()
-            ? "translateX"
-            : "translateY";
-        const offsetProperty: string = this.isHorizontal() ? "offsetLeft" : "offsetTop";
-        const prev: number = this.activeIndicatorRef[offsetProperty];
-        this.activeIndicatorRef.style[gridProperty] = `${this.activeTabIndex + 1}`;
-        const next: number = this.activeIndicatorRef[offsetProperty];
-        this.activeIndicatorRef.style[gridProperty] = `${this.prevActiveTabIndex + 1}`;
-        const dif: number = next - prev;
-        this.activeIndicatorRef.style.transform = `${translateProperty}(${dif}px)`;
-        this.activeIndicatorRef.classList.add("activeIndicatorTransition");
-        this.activeIndicatorRef.addEventListener("transitionend", () => {
-            this.ticking = false;
-            this.activeIndicatorRef.style[gridProperty] = `${this.activeTabIndex + 1}`;
-            this.activeIndicatorRef.style.transform = `${translateProperty}(0px)`;
-            this.activeIndicatorRef.classList.remove("activeIndicatorTransition");
-        });
-    }
 
     /**
      * The adjust method for FASTTabs