+++ packages/web-components/fast-foundation/src/toolbar/stories/toolbar.stories.ts
@@ -1,8 +1,8 @@
 import { html } from "@microsoft/fast-element";
-import { Orientation } from "@microsoft/fast-web-utilities";
 import type { Meta, Story, StoryArgs } from "../../__test__/helpers.js";
 import { renderComponent } from "../../__test__/helpers.js";
 import type { FASTToolbar } from "../toolbar.js";
+import { ToolbarOrientation } from "../toolbar.options.js";
 
 const storyTemplate = html<StoryArgs<FASTToolbar>>`
     <fast-toolbar
@@ -17,7 +17,7 @@ const storyTemplate = html<StoryArgs<FASTToolbar>>`
 export default {
     title: "Toolbar",
     argTypes: {
-        orientation: { control: "radio", options: Object.values(Orientation) },
+        orientation: { control: "radio", options: Object.values(ToolbarOrientation) },
         ariaLabel: { control: "text" },
         ariaLabelledby: { control: "text" },
         storyContent: { table: { disable: true } },
@@ -28,8 +28,6 @@ export const Toolbar: Story<FASTToolbar> = renderComponent(storyTemplate).bind({
 Toolbar.args = {
     storyContent: html`
         <button>Button</button>
-        <button slot="end">End Slot Button</button>
-        <button slot="start">Start Slot Button</button>
         <select>
             <option>Option 1</option>
             <option>Second option</option>
@@ -48,6 +46,33 @@ Toolbar.args = {
             Checkbox 3
         </label>
         <input type="text" name="text" id="text-input" />
+    `,
+};
+
+export const ToolbarWithSlottedStartEnd: Story<FASTToolbar> = Toolbar.bind({});
+ToolbarWithSlottedStartEnd.args = {
+    storyContent: html`
+        <svg slot="start" width="20" height="20"><use href="#test-icon" /></svg>
+        <button>Button</button>
+        <select>
+            <option>Option 1</option>
+            <option>Second option</option>
+            <option>Option 3</option>
+        </select>
+        <label for="check-1">
+            <input type="checkbox" name="checkbox" id="check-1" />
+            Checkbox 1
+        </label>
+        <label for="check-2">
+            <input type="checkbox" name="checkbox" id="check-2" />
+            Checkbox 2
+        </label>
+        <label for="check-3">
+            <input type="checkbox" name="checkbox" id="check-3" />
+            Checkbox 3
+        </label>
+        <input type="text" name="text" id="text-input" />
+        <svg slot="end" width="20" height="20"><use href="#test-icon-2" /></svg>
     `,
 };