+++ packages/web-components/fast-foundation/src/tree-item/tree-item.template.ts
@@ -1,5 +1,6 @@
 import { elements, ElementViewTemplate } from "@microsoft/fast-element";
 import { children, html, ref, slotted, when } from "@microsoft/fast-element";
+import { staticallyCompose } from "../utilities/template-helpers.js";
 import { endSlotTemplate, startSlotTemplate } from "../patterns/start-end.js";
 import type { FASTTreeItem, TreeItemOptions } from "./tree-item.js";
 
@@ -42,7 +43,7 @@ export function treeItemTemplate<T extends FASTTreeItem>(
                                 ${ref("expandCollapseButton")}
                             >
                                 <slot name="expand-collapse-glyph">
-                                    ${options.expandCollapseGlyph ?? ""}
+                                    ${staticallyCompose(options.expandCollapseGlyph)}
                                 </slot>
                             </div>
                         `