+++ packages/web-components/fast-ssr/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log - @microsoft/fast-ssr
 
-This log was last generated on Fri, 02 Dec 2022 01:18:22 GMT and should not be manually modified.
+This log was last generated on Tue, 28 Mar 2023 22:14:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.0-beta.27
+
+Tue, 28 Mar 2023 22:14:10 GMT
+
+### Changes
+
+- Bump @microsoft/fast-element to v2.0.0-beta.23
+- Bump @microsoft/fast-foundation to v3.0.0-alpha.27
+
+## 1.0.0-beta.26
+
+Sat, 11 Mar 2023 00:09:49 GMT
+
+### Changes
+
+- Bump @microsoft/fast-element to v2.0.0-beta.22
+- Bump @microsoft/fast-foundation to v3.0.0-alpha.26
+
+## 1.0.0-beta.25
+
+Tue, 14 Feb 2023 04:02:36 GMT
+
+### Changes
+
+- Bump @microsoft/fast-element to v2.0.0-beta.21
+- Bump @microsoft/fast-foundation to v3.0.0-alpha.25
+
+## 1.0.0-beta.24
+
+Wed, 11 Jan 2023 22:07:44 GMT
+
+### Changes
+
+- fix(fast-ssr): update to use html.partial (roeisenb@microsoft.com)
+- Bump @microsoft/fast-element to v2.0.0-beta.20
+- Bump @microsoft/fast-foundation to v3.0.0-alpha.24
 
 ## 1.0.0-beta.23